Re-organizing launch form components into a separate folder
