--- src/components/Launch/index.ts
@@ -1,2 +1,0 @@
-export * from './LaunchPlansTable';
-export * from './SchedulesTable';
